backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.0,<24.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
celery>=5.3,<6.0
redis>=5.0,<6.0

# Optional: For S3 storage
# boto3>=1.28,<2.0
# django-storages>=1.14,<2.0

# Development
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
black>=23.0,<24.0
flake8>=6.1,<7.0
